INode now contains the INode[] getParents() method, so that Gibbs' (it just now occurred to me that this could potentially be Gibb's...) Sampling can use the Markov Blanket in its calculations.
In other news, Gibbs' Sampling now actually uses the Markov blanket (thanks to a little help from the MIT slides). In the typical example with Sprinkler and Rain, it converged to approximately the same value as the other algorithms. Very exciting!

The only outstanding problem was an InputMismatchException in Dataset running sample2.data. Working to fix it now.