Added getChildren() and addChild(INode) to the INode interface so that the Markov blanket can be gathered efficiently. Correspondingly, Gibbs Sampling was updated to use the getChildren() method to gather the Markov blanket, signifying a notable increase in performance.
As a small change, Control now tests to see if a network has greater than a certain intractable number (in this case, 30) of nodes in the network. If the number of nodes exceeds this value, then Control automatically won't run the Enumeration Ask algorithm.

Redid package structure so that it won't break other packages that have similar names.